+++ common/src/main/java/com/clefal/teams/network/server/C2SAcceptApplicationPacket.java
@@ -1,12 +1,16 @@
 package com.clefal.teams.network.server;
 
 import com.clefal.nirvana_lib.network.C2SModPacket;
+import com.clefal.nirvana_lib.utils.NetworkUtils;
+import com.clefal.teams.network.client.S2CInvitationPacket;
 import com.clefal.teams.server.ATServerTeam;
 import com.clefal.teams.server.IHasTeam;
+import com.clefal.teams.server.Invitation;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.network.chat.Component;
 import net.minecraft.server.level.ServerPlayer;
 
+import java.util.Map;
 import java.util.UUID;
 
 public class C2SAcceptApplicationPacket implements C2SModPacket {
@@ -26,6 +30,7 @@ public class C2SAcceptApplicationPacket implements C2SModPacket {
             ATServerTeam team = ((IHasTeam) player).getTeam();
             ServerPlayer player1 = player.getServer().getPlayerList().getPlayer(uuid);
             if (player1 != null){
+                team.markApplicationAsRemoval(player1.getUUID());
                 team.addPlayer(player1);
             } else {
                 player.sendSystemMessage(Component.translatable("teams.error.no_player"));