+++ common/src/main/java/com/clefal/teams/server/Application.java
@@ -1,6 +1,7 @@
 package com.clefal.teams.server;
 
 import com.google.common.base.Objects;
+import net.minecraft.nbt.CompoundTag;
 
 import java.util.UUID;
 
@@ -21,5 +22,20 @@ public class Application extends ExpirableObject{
     @Override
     public int hashCode() {
         return Objects.hashCode(applicant);
+    }
+
+    public CompoundTag toNBT(){
+        CompoundTag compoundTag = new CompoundTag();
+        compoundTag.putUUID("uuid", applicant);
+        compoundTag.putInt("life", lifetime);
+        return compoundTag;
+    }
+
+    public static Application fromNBT(CompoundTag compoundTag){
+        UUID uuid = compoundTag.getUUID("uuid");
+        int life = compoundTag.getInt("life");
+        Application application = new Application(uuid);
+        application.lifetime = life;
+        return application;
     }
 }