+++ common/src/main/java/com/clefal/teams/utils/expirable/ScreenUtils.java
@@ -1,0 +1,12 @@
+package com.clefal.teams.utils.expirable;
+
+import lombok.experimental.UtilityClass;
+import net.minecraft.client.Minecraft;
+
+@UtilityClass
+public class ScreenUtils {
+
+    public void refreshIfIsTeamScreen(){
+        if (Minecraft.getInstance().screen instanceof )
+    }
+}